backend/requirements.txt
fastapi==0.110.0  # Updated for bug fixes
uvicorn==0.29.0  # Updated for performance
python-multipart==0.0.9
mlx-whisper==0.4.2  # This is crucial for Apple Silicon GPU acceleration
openai-whisper==20240930  # Provides the 'whisper' module used by mlx_whisper for model loading
python-dotenv==1.0.1
cryptography==42.0.5
slowapi==0.1.7 # For API rate limiting, if needed
tqdm==4.67.1 # For progress bar (used by WebSocketManager)
sentencepiece==0.2.0 # Often a dependency for tokenizers/model loading
python-docx==1.1.2 # If backend is responsible for generating .docx files
numpy==1.26.4  # Updated for better Apple Silicon support
sounddevice==0.5.2
webrtcvad==2.0.10
